+++ ubuntu-apache2-phpfpm-dev-7.0/startup-php-sessions.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+if [[ ! -d /data/php-sessions ]] ; then
+    mkdir /data/php-sessions
+    chown -R www-data:www-data /data/php-sessions
+fi